Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

I have tested it as far as possible.
Got andro-SYR (XDA) to tell me what characters where on the keyboard.
He was able to identify all the basic characters, but most characters where you shall use "shift" to get to, are still missing.
Hopefully someone with an Arabic Mini Pro or Pro will be able to help, to get the rest of the characters working.

Patch-set: 1