Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

The non-standard symlink command does not exist. Using the standard ln command, makes this work.
Have tested this and now it is for example possible to take screenshots and use other parts of the system, that tries to save data to /sdcard or /mnt/sdcard .

Patch-set: 1