Update patch set 1
Patch Set 1:

So you do not wish the symbolic links to get created?
At the moment you are using a non-standard 'symlink' command, that does not exist on the phone. This results in '/sdcard' and '/mnt/sdcard' not getting created.
A lot of system apps depend on '/sdcard', so they fail without this fix (or if you wish to make an alias from 'ln -s' to 'symlink').

Patch-set: 1