Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

the init.*.rc and init.rc parser is NOT shell commands.  It is not a shell script.

symlink is a valid init.rc command.  By replacing symlink with 'ln -s' you are just ignoring those two lines (no symlinks are being created at all).  This may be causing some other behavior, but this is not right

Please see this file for a list of valid commands:
 
https://github.com/CyanogenMod/android_system_core/blob/ics/init/init_parser.c

Patch-set: 1